rename recordingStatus component as recordingStatusText
